Add user/post-data association
A logged in user can create a post that will be tied to their account. Their username will also be displayed as a heading and link on top of each post. Users can also comment on posts which will also be tied to their account and displayed along with the comment.